fixed signal generation

+++ src/THREADS/DATA_INTAKE/mockSignal.c
@@ -5,9 +5,11 @@ void mockSignal(float * signal, const size_t size, const float amplitude, const 
     float t;
 
     for (int i = 0; i < size; i++) {
-        t = i / SAMPLING_FREQ;
+        t = (float)i / SAMPLING_FREQ;
         float sine_value = amplitude * sin(2 * PI * frequency * t);
+
         signal[i] = sine_value;
+        
     }
 
 }
